backend/requirements.txt
# Core Dependencies
deepagents>=0.1.0
langgraph>=0.2.0
langchain>=0.1.0
openai>=1.0.0

# Data Processing
pandas>=2.0.0
dask[complete]>=2024.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# Database & Caching
redis>=5.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Visualization & Analysis
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1